Transferred files to Documents
Library, temporary, cache and config.json files has been transferred to user Documents folder.

+++ functions/folder.py
@@ -1,4 +1,5 @@
 from glob import glob
+from tkinter.messagebox import askyesno
 import customtkinter as ctk
 import subprocess
 import os
@@ -10,50 +11,12 @@ def open_folder(path: str):
     subprocess.run([FILEBROWSER_PATH, path])
 
 
-def clear_folder(path: str, app: ctk.CTk):
-    warning = ctk.CTkToplevel(master=app)
-    warning.title(f"Delete {path} wallpapers!")
-    warning.iconbitmap("assets/colored_main.ico")
-
-    window_height = 150
-    window_width = 450
-
-    screen_width = app.winfo_screenwidth()
-    screen_height = app.winfo_screenheight()
-
-    x = int((screen_width / 2) - (window_width / 2))
-    y = int((screen_height / 2) - (window_height / 2))
-
-    warning.geometry(f"{window_width}x{window_height}+{x}+{y}")
-
-    warning.grid_rowconfigure(0, weight=1)
-    warning.grid_columnconfigure(0, weight=1)
-
-    ctk.CTkLabel(
-        master=warning,
-        text="You image will be deleted permanently. You can't restore them later!",
-    ).grid(row=0, columnspan=3, padx=20, pady=20, sticky="nwse")
-
-    def delete_files():
+def clear_folder(path: str, name: str, app: ctk.CTk):
+    confirm = askyesno(
+        f"Clear {name} files",
+        f"All files in the {name} folder will be permanently delete. Do you want to delete them?",
+    )
+    if confirm:
         files = glob(f"{path}/*.jpg") + glob(f"{path}/*.png") + glob(f"{path}/*.txt")
         for file in files:
             os.remove(file)
-        warning.destroy()
-
-    ctk.CTkButton(
-        master=warning,
-        text="Delete",
-        fg_color=["#a44", "#800"],
-        hover_color=["#800", "#a44"],
-        command=delete_files,
-    ).grid(row=1, column=1, padx=10, pady=10, sticky="e")
-
-    ctk.CTkButton(
-        master=warning,
-        text="Cancel",
-        fg_color=["#aaa", "#777"],
-        hover_color=["#777", "#aaa"],
-        command=warning.destroy,
-    ).grid(row=1, column=2, padx=10, pady=10, sticky="e")
-
-    warning.mainloop()
